+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [8.29.1] - 2019-04-08
+
+### Fixed
+
+- **DatePicker** input now respects parent width instead of allways being 199px wide.
+
+### Added
+
+- **EXPERIMENTAL_Filter** as Experimental component.
+- **EXPERIMENTAL_Conditions** add `onChangeObjectCallback` prop so it can be passed to the object atom as a callback.
+- **Table** add FilterBar option in toolbar and example with FilterBar.
+
 ## [8.29.0] - 2019-04-08
 
 ### Added